+++ spec/spec_helper.rb
@@ -1,0 +1,11 @@
+require 'chefspec'
+require 'chefspec/berkshelf'
+require 'pry'
+
+RSpec.configure do |config|
+  # Specify the operating platform to mock Ohai data from (default: nil)
+  config.platform = 'ubuntu'
+
+  # Specify the operating version to mock Ohai data from (default: nil)
+  config.version = '12.04'
+end