+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file is used to list changes made in each version of pantry cookbook.
 
 ## 0.3.6
 * Added rbenv
+* Save credentials for gem source
 
 ## 0.3.5
 * Moved database & user creation in separate recipe